added btc api test
